Create destroy_the_dots.py
This program is a simple game using the guizero library in Python. The objective of the game is to click on red dots that appear on a grid. Each time a red dot is clicked, the score increases. The game speeds up as the score increases, and it ends when all cells turn red